adaptipy/requirements.txt
Django==5.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-cors-headers==4.3.1
requests==2.32.5
python-dotenv==1.1.1
openai==2.5.0
sentence-transformers==5.1.1
scikit-learn==1.7.2
numpy==2.2.6